+++ ErsatzTV.Application/Channels/Commands/UpdateChannel.cs
@@ -14,6 +14,7 @@ namespace ErsatzTV.Application.Channels.Commands
         string Logo,
         string PreferredLanguageCode,
         StreamingMode StreamingMode,
+        string Watermark,
         ChannelWatermarkMode WatermarkMode,
         ChannelWatermarkLocation WatermarkLocation,
         ChannelWatermarkSize WatermarkSize,