+++ ErsatzTV.Core/Domain/Metadata/ArtworkKind.cs
@@ -5,6 +5,7 @@
         Poster = 0,
         Thumbnail = 1,
         Logo = 2,
-        FanArt = 3
+        FanArt = 3,
+        Watermark = 4
     }
 }